spatialdiva/preprocessing/requirements.txt
Updating dependencies
Resolving dependencies...

Writing lock file
anndata==0.10.9 ; python_version >= "3.10" and python_version < "4.0"
array-api-compat==1.10.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.56.0 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2025.2.0 ; python_version >= "3.10" and python_version < "4.0"
h5py==3.13.0 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.23.5 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
joblib==1.4.2 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.8 ; python_version >= "3.10" and python_version < "4.0"
legacy-api-wrap==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
llvmlite==0.44.0 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.10.0 ; python_version >= "3.10" and python_version < "4.0"
natsort==8.4.0 ; python_version >= "3.10" and python_version < "4.0"
networkx==3.4.2 ; python_version >= "3.10" and python_version < "4.0"
numba==0.61.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
patsy==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "4.0"
pynndescent==0.5.13 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
scanpy==1.11.0 ; python_version >= "3.10" and python_version < "4.0"
scikit-learn==1.5.2 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.15.2 ; python_version >= "3.10" and python_version < "4.0"
seaborn==0.13.2 ; python_version >= "3.10" and python_version < "4.0"
session-info2==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
statsmodels==0.14.4 ; python_version >= "3.10" and python_version < "4.0"
threadpoolctl==3.5.0 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0"
umap-learn==0.5.7 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
